refactor: adjust color file location
